+ Ajout : PhpStan #24

+++ src/Utils/StringHelper.php
@@ -35,6 +35,6 @@ class StringHelper
             return '';
         }
 
-        return substr(str_shuffle(str_repeat($x = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ', ceil($length / \strlen($x)))), 1, $length);
+        return substr(str_shuffle(str_repeat($x = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ', (int) ceil($length / \strlen($x)))), 1, $length);
     }
 }
